python-livekit-speech-to-text--agent/requirements.txt
livekit-agents
livekit-plugins-assemblyai
fastapi
uvicorn
python-dotenv
